Changed reward structure: previously both BORDER_PENALTY and WIN_REWARD were 300. #This the agent to jump over a border near the end point due to less number of steps required #Hence, increased BORDER_PENALTY and reduced WIN_REWARD
Also, changed logic so that agent couldn't get WIN_REWARD when it jumped 
over the border

Spawning and end locations were previously incorrect. Because math is 
hard!!

Also, as the episode didn't end on reaching the destination, the agent 
kept switching back and forth accumulating the WIN_REWARD.

Changed color of agent to green when it reaches WIN location.